Prevent hang when the point is at the bottom of the buffer
If the point is at the bottom of the buffer and which-function-mode
was enabled, Emacs freezed. The reason was that the code entered the
"unmatched parens" part of the logic and got stuck navigating blocks
forward and backward without ever matching a block correctly.

This can be fixed if we check for the end of buffer case explicitly
and return a whole buffer as a block. This makes a point "in between
functions" return the same function information as a point at the end
of the buffer. Graphically:

protocol P { func f() }
extension P { func g() { f() } }
struct X<T> : P {
    func f() { print("unconstrained") }
}
<--- Point here (which-function-mode returns "P")
struct X<T> : where T : BinaryInteger {
    func f() { print("constrained")
}
<--- Point here (which-function-mode returns "P")